Refactor: enum type 들을 type 패키지로 묶음
