+++ src/test/java/com/zero/triptalk/planner/service/PlannerServiceTest.java
@@ -1,9 +1,9 @@
 package com.zero.triptalk.planner.service;
 
 import com.zero.triptalk.planner.dto.PlannerRequest;
-import com.zero.triptalk.planner.dto.PlannerStatus;
 import com.zero.triptalk.planner.entity.Planner;
 import com.zero.triptalk.planner.repository.PlannerRepository;
+import com.zero.triptalk.planner.type.PlannerStatus;
 import com.zero.triptalk.user.entity.UserEntity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -64,6 +64,6 @@ class PlannerServiceTest {
         //when
         plannerService.deletePlanner(plannerId);
         //then
-        verify(plannerRepository,times(1)).deleteById(plannerId);
+        verify(plannerRepository, times(1)).deleteById(plannerId);
     }
 }
